requirements.txt
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.8.3
cffi==2.0.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==45.0.7
daphne==4.1.2
Django==5.2.5
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
isort==6.0.1
msgpack==1.1.1
pillow==11.3.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.10.0
zope.interface==7.2
